Add support for Scala 2.13
Daffodil 4.0.0 and later will use Scala 2.13--this modifies the plugin
set the scalaVersion setting to 2.13 for newer versions of Daffodil. It
also adds support for Scala 2.13 minimum supported JDK versions like we
have for Scala 2.12.

This also adds support for packgeDaffodilBinVersions 4.0.0+. Instead of
using double %% for packageDaffodilBinVersions libraryDependencies, we
now use a single % and manullay add the _2.12 or _2.13 depending on the
Daffodil versions. This ensures packageDaffodilBin uses the right
version of Scala for the right Daffodil version.

This also remove JavaConverters from the DaffodilSaver file. This
plugin, and thus this file is compiled with Scala 2.12, but Scala 2.13
will be on the classpath when DaffodilSaver is forked when building
saved parsers for newer Daffodil versions. Fortunately, DaffodilSaver is
pretty simply and only has the one place that really relies on Scala
2.12, which is easy enough to remove by using Java forEach instead of
asScala. If more 2.13 incompatibilities start to crop up, we may want to
rewrite DaffodilSaver in Java to avoid Scala dependency issues.

Closes #95