Simplify setting scalaVersion
The way we currently set scalaVersionis complex--we try to use the
version of scala that was distributed with daffodilVersion, but will
potentially upgrade to new scala versions depending on the JDK version.

But all this is pretty unnecessary. Scala patch updates guarantee binary
compatability, so we can just always use the latest scala patch version
that is compatable with daffodilVersion and things shoudl always work.
This ensures that users of the plugin always use the latest Scala
version, including bug fixes and JDK compatability.

This means we no longer need to maintain JDK compatability versions or a
full mapping of all Daffodil version to Scala version.

Note that this also uses ModuleId.version, which seems to help
scala-steward recognize that these are scala versions and will auto
update them when new versions are relased.

Also uses a more simplifed syntax for disabling scala-steward for the
daffodil dependencies.

Also updates all scala versions to their latest available patch
versions.

Closes #143